Merged 2,375 commits from upstream flutter/packages (July 2024 → Dec 2025)

Key changes:
- Updated local_auth to v3.0.0 (structured error handling)
- Gradle 9 compatibility
- Pigeon 26 updates
- Java 17 migration
- Numerous dependency updates

Custom modifications preserved:
- STRONG-only biometric enforcement on Android 11+ (API 30+)
- SDK compatibility with Dart 3.2.3 / Flutter 3.16.6
- Git dependency reference for local_auth_android

Conflicts resolved in:
- packages/local_auth/local_auth/pubspec.yaml
- packages/local_auth/local_auth_android/pubspec.yaml
- packages/local_auth/local_auth_android/android/src/main/java/io/flutter/plugins/localauth/AuthenticationHelper.java

Security enhancement verified: Build.VERSION.SDK_INT check for STRONG authenticators still in place.

+++ packages/pigeon/platform_tests/test_plugin/android/src/test/kotlin/com/example/test_plugin/PrimitiveTest.kt
@@ -1,4 +1,4 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -21,7 +21,7 @@ class PrimitiveTest {
     val binaryMessenger = mockk<BinaryMessenger>(relaxed = true)
     val api = mockk<PrimitiveHostApi>(relaxed = true)
 
-    val input = 1
+    val input = 1L
 
     val channelName = "dev.flutter.pigeon.pigeon_integration_tests.PrimitiveHostApi.anInt"
     val handlerSlot = slot<BinaryMessenger.BinaryMessageHandler>()
@@ -38,7 +38,7 @@ class PrimitiveTest {
       it?.rewind()
       @Suppress("UNCHECKED_CAST") val wrapped = codec.decodeMessage(it) as List<Any>?
       assertNotNull(wrapped)
-      wrapped?.let { assertEquals(input.toLong(), wrapped[0]) }
+      wrapped?.let { assertEquals(input, wrapped[0]) }
     }
 
     verify { binaryMessenger.setMessageHandler(channelName, handlerSlot.captured) }
